requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
whitenoise==5.1.0
django-heroku==0.3.1
dj-database-url==0.5.0
